# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BilibiliComment Admin API
  version: 0.1.0
  description: 单管理员任务管理、评论观测与系统状态 API。所有 bigint ID 均按字符串传输。
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/tasks/{taskId}/executions:
    parameters:
      - {$ref: '#/components/parameters/TaskId'}
    get:
      tags: [Executions]
      operationId: listTaskExecutions
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: 执行历史分页
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExecutionPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/executions/{executionId}:
    get:
      tags: [Executions]
      operationId: getExecution
      parameters:
        - {$ref: '#/components/parameters/ExecutionId'}
      responses:
        '200':
          description: 单次执行详情
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskExecution'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/tasks/{taskId}/comments:
    parameters:
      - {$ref: '#/components/parameters/TaskId'}
    get:
      tags: [Executions]
      operationId: listTaskComments
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: 当前任务独立评论表的游标分页
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommentPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PositiveId'}
    ExecutionId:
      name: executionId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PositiveId'}
...
    Cursor:
      name: cursor
      in: query
      description: 服务端返回的不透明分页边界；客户端必须原样回传，不得解析或自行构造。
      schema: {type: string, minLength: 1}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
...
  responses:
    BadRequest:
      description: 请求格式错误
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Unauthorized:
      description: 未登录或凭据错误
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
...
    NotFound:
      description: 资源不存在
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
...
    Unprocessable:
      description: 参数验证失败
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
...
  schemas:
    PositiveId: {type: string, pattern: '^[1-9][0-9]*$'}
...
    ExecutionStatus: {type: string, enum: [QUEUED, RUNNING, RETRY_WAIT, SUCCEEDED, FAILED, CANCELLED]}
...
    ExecutionPhase:
      type: [string, 'null']
      enum: [RESOLVING_SOURCE, FETCHING_COMMENTS, PERSISTING_COMMENTS, WAITING_RATE_LIMIT, SCHEDULING_NEXT, null]
...
    TaskExecution:
      type: object
      required: [id, taskId, trigger, status, phase, pagesFetched, commentsDiscovered, commentsInserted, duplicatesSkipped, retryCount, cursor, errorSummary, startedAt, finishedAt, lastHeartbeatAt]
      properties:
        id: {$ref: '#/components/schemas/PositiveId'}
        taskId: {$ref: '#/components/schemas/PositiveId'}
        trigger: {type: string, enum: [SCHEDULED, MANUAL, DISCOVERED]}
        status: {$ref: '#/components/schemas/ExecutionStatus'}
        phase: {$ref: '#/components/schemas/ExecutionPhase'}
        pagesFetched: {type: integer, minimum: 0}
        commentsDiscovered: {type: integer, minimum: 0}
        commentsInserted: {type: integer, minimum: 0}
        duplicatesSkipped: {type: integer, minimum: 0}
        retryCount: {type: integer, minimum: 0}
        cursor: {type: [string, 'null']}
        errorSummary: {type: [string, 'null']}
        startedAt: {type: [string, 'null'], format: date-time}
        finishedAt: {type: [string, 'null'], format: date-time}
        lastHeartbeatAt: {type: [string, 'null'], format: date-time}
    Comment:
      type: object
      required: [id, rpid, parentRpid, mid, uname, avatar, currentLevel, content, ctime]
      properties:
        id: {$ref: '#/components/schemas/PositiveId'}
        rpid: {$ref: '#/components/schemas/PositiveId'}
        parentRpid: {oneOf: [{$ref: '#/components/schemas/PositiveId'}, {type: 'null'}]}
        mid: {type: string}
        uname: {type: [string, 'null']}
        avatar: {type: [string, 'null']}
        currentLevel: {type: [integer, 'null']}
        content: {type: [string, 'null']}
        ctime: {type: string, format: date-time}
...
    ExecutionPage:
      type: object
      required: [items, nextCursor, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/TaskExecution'}}
        nextCursor: {type: [string, 'null']}
        total: {type: integer, minimum: 0}
    CommentPage:
      type: object
      required: [items, nextCursor, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Comment'}}
        nextCursor: {type: [string, 'null']}
        total: {type: integer, minimum: 0}
    ApiProblem:
      type: object
      required: [status, code, detail]
      properties:
        type: {type: [string, 'null'], format: uri}
        title: {type: [string, 'null']}
        status: {type: integer, minimum: 400, maximum: 599}
        code: {type: string}
        detail: {type: string}
        traceId: {type: [string, 'null']}
        fieldErrors: {type: object, additionalProperties: {type: string}}
        existingTaskId: {oneOf: [{$ref: '#/components/schemas/PositiveId'}, {type: 'null'}]}